goci-tools/goci-data-release-qc/src/main/java/uk/ac/ebi/spot/goci/service/SolrQueryService.java
package uk.ac.ebi.spot.goci.service;

import org.apache.http.HttpEntity;
import org.apache.http.HttpHost;
import org.apache.http.client.config.RequestConfig;
import org.apache.http.client.methods.CloseableHttpResponse;
import org.apache.http.client.methods.HttpGet;
import org.apache.http.impl.client.CloseableHttpClient;
import org.apache.http.impl.client.HttpClients;
import org.apache.http.util.EntityUtils;
import org.slf4j.Logger;
import org.slf4j.LoggerFactory;
import org.springframework.beans.factory.annotation.Value;
import org.springframework.core.io.Resource;
import org.springframework.stereotype.Service;

import java.io.BufferedReader;
import java.io.IOException;
import java.io.InputStreamReader;
import java.util.Properties;

/**
 * Created by dwelter on 29/06/16.
 */
@Service
public class SolrQueryService {

    private Logger log = LoggerFactory.getLogger(getClass());

    protected Logger getLog() {
        return log;
    }

    @Value("${catalog.stats.file}")
    private Resource catalogStatsFile;


    public SolrQueryService(){

    }

    public String getLastReleaseDate(){
        String releasedate;

        Properties properties = new Properties();
        try {
            properties.load(catalogStatsFile.getInputStream());
            releasedate = properties.getProperty("releasedate");

        }
        catch (IOException e) {
            throw new RuntimeException(
                    "Unable to return catolog stats: failed to read catalog.stats.file resource", e);
        }
        return releasedate;
    }

    public String buildSolrQuery(String releaseDate){
        return "bar";
    }

    public void querySolr(String searchString) throws IOException {

        System.out.println(searchString);
        CloseableHttpClient httpclient = HttpClients.createDefault();
        HttpGet httpGet = new HttpGet(searchString);
        if (System.getProperty("http.proxyHost") != null) {
            HttpHost proxy;
            if (System.getProperty("http.proxyPort") != null) {
                proxy = new HttpHost(System.getProperty("http.proxyHost"), Integer.parseInt(System.getProperty
                        ("http.proxyPort")));
            }
            else {
                proxy = new HttpHost(System.getProperty("http.proxyHost"));
            }
            httpGet.setConfig(RequestConfig.custom().setProxy(proxy).build());
        }

        String file = null;
        try (CloseableHttpResponse response = httpclient.execute(httpGet)) {
            getLog().debug("Received HTTP response: " + response.getStatusLine().toString());
            HttpEntity entity = response.getEntity();

            BufferedReader br = new BufferedReader(new InputStreamReader(entity.getContent()));

            String output;
            while ((output = br.readLine()) != null) {


            }

            EntityUtils.consume(entity);
        }
        if (file == null) {

            //TO DO throw exception here and add error handler
            file =
                    "Some error occurred during your request. Please try again or contact the GWAS Catalog team for assistance";
        }


    }


    public Resource getCatalogStatsFile(){return catalogStatsFile;};

    public void setCatalogStatsFile(Resource csf){
        this.catalogStatsFile = csf;
    }
}
